+++ src/Frontend/code/pages/play/lobby.ts
@@ -2,8 +2,8 @@ import { router } from "../../routes/router";
 import { flashButton, getButton, getTable, toggleButton } from "./utils/stylingComponents";
 import { getLobbyOptionsHTML } from "./utils/concreteComponents";
 import { TLobby } from "./lobbyTyping";
-import { TDynamicLobbySettings } from "./lobbyTyping";
-import { LobbyLogic } from "./lobbyLogic";
+import { TDynamicLobbySettings, TMapType, TMatchDuration, TMatchMode } from "./lobbyTyping";
+//import { LobbyLogic } from "./lobbyLogic";
 import { lobby } from "../../services/LobbyService";
 import { ROLES, SIDES } from "../../match/matchSharedDependencies/sharedTypes";
 
@@ -78,8 +78,7 @@ export const LobbyPage = {
         const formChangeSettings = document.getElementById('settings-change-form') as HTMLElement;
         formChangeSettings.addEventListener('submit', async (e) => {
             e.preventDefault()
-            await LobbyLogic.fetchAndSubmitSettings();
-            this.renderSettings();
+            this.fetchAndSubmitSettings();
         })
     },
 
@@ -203,7 +202,7 @@ export const LobbyPage = {
         const closeDialogButton = document.getElementById("btn-close-dialog") as HTMLElement;
         closeDialogButton.addEventListener("click", (e) => {
             e.preventDefault();
-            LobbyLogic.fetchAndAddPlayerToSlot(settingsDialog, team, role);
+            this.fetchAndAddPlayerToSlot(settingsDialog, team, role);
             this.renderSlots();
             settingsDialog.close();
             settingsDialog.remove();
@@ -258,5 +257,37 @@ export const LobbyPage = {
         joinWithdrawButton.classList.add("w-full");
 
         participantsElement.appendChild(joinWithdrawButton)
+    },
+
+    //Settings logic
+    fetchAndSubmitSettings() {
+        const map = (document.getElementById('match-map') as HTMLSelectElement).value;
+        const mode = (document.getElementById('match-mode') as HTMLSelectElement).value;
+        const duration = (document.getElementById('match-duration') as HTMLSelectElement).value;
+        lobby.updateSettingsIN({
+            map: map as TMapType,
+            mode: mode as TMatchMode,
+            duration: duration as TMatchDuration
+        })
+        console.log("New settings applied!")
+    },
+
+    fetchAndAddPlayerToSlot(settingsDialog: HTMLDialogElement,
+            team: SIDES, role: ROLES) {
+        const form = settingsDialog.querySelector("form") as HTMLFormElement;
+        if (!form.reportValidity()) return;
+        const aliasInput = document.getElementById("player-alias") as HTMLInputElement;
+        const alias = aliasInput.value;
+        const spriteIdInput = document.getElementById("player-paddle") as HTMLInputElement;
+        const spriteID = Number(spriteIdInput.value)
+        lobby.addFriendlyPlayerIN({
+            id: -1, //This is supposed to be generated by backend
+            nickname: alias,
+            spriteID: spriteID,
+            team: team,
+            role: role 
+        });
+        console.log(`Player saved with alias ${alias} and sprite id ${spriteID}`)
+        console.log(`Player added to team ${team} and role ${role}!`)
     },
 }
