+++ src/Frontend/code/pages/play/utils/helpers.ts
@@ -1,5 +1,19 @@
-import { TSlots } from "../lobbyLogic";
 import { TMapType } from "../lobbyTyping";
+import { SIDES, ROLES } from "../../../match/matchSharedDependencies/sharedTypes";
+
+type TPlayerInSlot = {
+    id: number,
+    spriteID: number
+}
+
+type TTeam = {
+    -readonly [key in keyof typeof ROLES]?: TPlayerInSlot | null //The shittiest fucking language I have ever seen in my life
+}
+
+export type TSlots = {
+    -readonly [key in keyof typeof SIDES]?: TTeam
+}
+
 
 export function getSlotsFromMap(map: TMapType): TSlots {
     const [amount, type, _size] = map.split("-");
